refactor  navbar for better reuse
lets left and right side children be added using a neat trick found at
https://medium.com/@martin_hotell/react-children-composition-patterns-with-typescript-56dfc8923c64
