+++ src/components/Navbar.tsx
@@ -1,10 +1,8 @@
-import React from "react";
+import React, { ReactNode } from "react";
 
 import tw from "tailwind.macro";
 
 import logo from "./bug_logo.png";
-import { ReactComponent as LinkSVG } from "./link.svg";
-import { ReactComponent as SaveSVG } from "./save-disk.svg";
 
 const Nav = tw.nav`
   flex items-center justify-between flex-wrap bg-gray-700 p-2
@@ -40,42 +38,25 @@ const NavButton = tw.button`
   inline-block text-sm px-4 py-2 mt-0 ml-1 leading-none border rounded text-gray-300 border-gray-300 hover:border-transparent hover:text-green-500 hover:bg-white
 `;
 
-const SaveIcon = tw(SaveSVG)`
-  fill-current inline-block h-4 w-4
-`;
-
-const LinkIcon = tw(LinkSVG)`
-  fill-current inline-block h-4 w-4
-`;
-
 interface NavbarProps {
-  onClick: any;
+  children?: {
+    left?: ReactNode;
+    right?: ReactNode;
+  };
 }
 
-const Navbar: React.FC<NavbarProps> = ({ onClick }) => (
+const Navbar: React.FC<NavbarProps> = ({ children }) => (
   <Nav>
     <NavHeader>
       <NavLogo src={logo} alt="transientBug logo" />
       <NavTitle>transientBug</NavTitle>
     </NavHeader>
     <NavItems>
-      <NavLeft>
-        {/* <NavLink
-          href="#responsive-header"
-        >
-          Docs
-        </NavLink> */}
-      </NavLeft>
-      <NavRight>
-        <NavButton onClick={() => onClick("save")} color="red">
-          <SaveIcon />
-        </NavButton>
-        <NavButton onClick={() => onClick("open-tb")}>
-          <LinkIcon />
-        </NavButton>
-      </NavRight>
+      <NavLeft>{children && children.left}</NavLeft>
+      <NavRight>{children && children.right}</NavRight>
     </NavItems>
   </Nav>
 );
 
 export default Navbar;
+export { NavButton };
